The project structure was changed due to requirements. All dependencies were reestablished.
