+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import registerServiceWorker from './registerServiceWorker';
+import App from './js/App';
+import registerServiceWorker from './js/workers/registerServiceWorker';
+import './styles/body.css';
 
 ReactDOM.render(<App />, document.getElementById('root'));
 registerServiceWorker();
