+++ src/js/common/components/sideNavigation/navigationItem.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Button from '../common/button';
+import Button from '../controls/button';
 import { Link } from 'react-router-dom';
 
 function NavigationItem(props) {
