+++ src/js/common/components/sideNavigation/sideNavigation.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import NavigationItem from './navigationItem';
-import '../../style/body/side-navigation/side-navigation.css';
+import '../../../../styles/flex-box/side-navigation/side-navigation.css';
 
 function SideNavigation(props) {
 	const linkTitlePairs = props.navigationItemsMeta;
