+++ src/js/pages/mySurveysPage/containers/mySurveysPage.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+function MySurveysPage() {
+	return (
+		<div className="my-surveys-page">
+			
+		</div>
+	);
+}
+
+export default MySurveysPage
