+++ src/js/pages/mySurveysPage/containers/newSurveyPage.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+function NewSurveyPage() {
+	return (
+		<div className="new-survey-page">
+			
+		</div>
+	);
+}
+
+export default NewSurveyPage
