+++ src/js/pages/newSurveyPage/containers/surveyTemplatesPage.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+function SurveyTemplatesPage() {
+	return (
+		<div className="survey-template-page">
+			
+		</div>
+	);
+}
+
+export default SurveyTemplatesPage
