+++ src/js/workers/registerServiceWorker.js
