+++ src/styles/body.css
@@ -1,0 +1,3 @@
+body {
+	margin: 0;
+}
